pyproject.toml
[project]
name = "pdfoptimizer"
version = "4.0.0"
description = "A powerful PDF utility that supports PDF compression, merging, splitting, image conversion, text-to-curves conversion, and bookmark management."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.0.0",
    "pikepdf>=8.0.0",
    "PyMuPDF>=1.23.0",
    "pyinstaller>=6.14.2",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "mistralai>=0.1.4",
    "python-datauri>=1.0.0",
    "python-docx>=0.8.11"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.0.0",
    "black>=24.4.2",
    "flake8>=7.0.0",
    "isort>=5.13.2",
    "mypy>=1.10.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
pdfoptimizer = "main:main"

[tool.setuptools]
packages = ["core", "ui"]

[tool.black]
line-length = 99
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 99
extend-ignore = "E203, W503"
exclude = [".venv", "venv", ".git", "__pycache__", "dist", "build"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["venv", ".venv"]
